backend/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow